fix(api): :ambulance: Correction du GET /partenaires
En HTTP 500 lorsque que la recherche n'était pas faire avec des critères de geo localisation

fix #64

+++ src/main/java/nc/opt/mobile/api/mobilis/partenaires/PartenaireRepository.java
@@ -32,7 +32,7 @@ public interface PartenaireRepository extends JpaRepository<Partenaire, Long> {
     )
     List<Partenaire> find(
         @Param("point") String point,
-        @Param("distance") long distance,
+        @Param("distance") Integer distance,
         @Param("q") String q,
         @Param("ville") String ville,
         @Param("codePostal") String codePostal,
